+++ src/de/mod.rs
@@ -1062,7 +1062,7 @@ mod tests {
                 de.write,
                 vec![
                     Start(BytesStart::borrowed_name(b"inner")),
-                    Text(BytesText::from_escaped_str("text")),
+                    Text(BytesText::from_escaped("text")),
                     Start(BytesStart::borrowed_name(b"inner")),
                     End(BytesEnd::borrowed(b"inner")),
                     End(BytesEnd::borrowed(b"inner")),
@@ -1099,7 +1099,7 @@ mod tests {
                 de.read,
                 vec![
                     Start(BytesStart::borrowed_name(b"inner")),
-                    Text(BytesText::from_escaped_str("text")),
+                    Text(BytesText::from_escaped("text")),
                     Start(BytesStart::borrowed_name(b"inner")),
                     End(BytesEnd::borrowed(b"inner")),
                     End(BytesEnd::borrowed(b"inner")),
@@ -1126,7 +1126,7 @@ mod tests {
                 vec![
                     // This comment here to keep the same formatting of both arrays
                     // otherwise rustfmt suggest one-line it
-                    Text(BytesText::from_escaped_str("text")),
+                    Text(BytesText::from_escaped("text")),
                 ]
             );
 
@@ -1149,15 +1149,12 @@ mod tests {
             assert_eq!(
                 de.read,
                 vec![
-                    Text(BytesText::from_escaped_str("text")),
+                    Text(BytesText::from_escaped("text")),
                     End(BytesEnd::borrowed(b"inner")),
                 ]
             );
             assert_eq!(de.write, vec![]);
-            assert_eq!(
-                de.next().unwrap(),
-                Text(BytesText::from_escaped_str("text"))
-            );
+            assert_eq!(de.next().unwrap(), Text(BytesText::from_escaped("text")));
             assert_eq!(de.next().unwrap(), End(BytesEnd::borrowed(b"inner")));
             assert_eq!(
                 de.next().unwrap(),
@@ -1200,7 +1197,7 @@ mod tests {
                 de.write,
                 vec![
                     Start(BytesStart::borrowed_name(b"skip")),
-                    Text(BytesText::from_escaped_str("text")),
+                    Text(BytesText::from_escaped("text")),
                     Start(BytesStart::borrowed_name(b"skip")),
                     End(BytesEnd::borrowed(b"skip")),
                     End(BytesEnd::borrowed(b"skip")),
@@ -1224,7 +1221,7 @@ mod tests {
                 de.write,
                 vec![
                     Start(BytesStart::borrowed_name(b"skip")),
-                    Text(BytesText::from_escaped_str("text")),
+                    Text(BytesText::from_escaped("text")),
                     Start(BytesStart::borrowed_name(b"skip")),
                     End(BytesEnd::borrowed(b"skip")),
                     End(BytesEnd::borrowed(b"skip")),
@@ -1246,7 +1243,7 @@ mod tests {
                 de.read,
                 vec![
                     Start(BytesStart::borrowed_name(b"skip")),
-                    Text(BytesText::from_escaped_str("text")),
+                    Text(BytesText::from_escaped("text")),
                     Start(BytesStart::borrowed_name(b"skip")),
                     End(BytesEnd::borrowed(b"skip")),
                     End(BytesEnd::borrowed(b"skip")),
@@ -1409,7 +1406,7 @@ mod tests {
                     br#"item name="hello" source="world.rs""#,
                     4
                 )),
-                Text(BytesText::from_escaped(b"Some text".as_ref())),
+                Text(BytesText::from_escaped("Some text")),
                 End(BytesEnd::borrowed(b"item")),
                 Start(BytesStart::borrowed(b"item2", 5)),
                 End(BytesEnd::borrowed(b"item2")),